Expose public
