+++ now.json
@@ -1,0 +1,5 @@
+{
+    "alias": "shortnim.me",
+    "cleanUrls": true,
+    "trailingSlash": false
+}
