+++ package.json
@@ -1,12 +1,11 @@
 {
   "name": "shortnim-server",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Server",
   "main": "main.js",
   "scripts": {
-    "test": "nodemon main.js",
     "start": "node main.js",
-    "dev": "now remove shortnim-server --safe -y && now --public && now alias"
+    "dev": "nodemon main.js"
   },
   "repository": {
     "type": "git",
@@ -23,8 +22,11 @@
   },
   "homepage": "https://github.com/Albermonte/nimiq-shortener#readme",
   "dependencies": {
-    "express": "^4.16.3",
-    "node-fetch": "^2.2.0",
-    "socket.io": "^2.1.1"
+    "express": "^4.17.1",
+    "node-fetch": "^2.6.0",
+    "socket.io": "^2.3.0"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.3"
   }
 }
