+++ public/redirect.js
