+++ public/statistics.html
@@ -76,6 +76,11 @@
     </style>
 </head>
 <header>
+    <noscript>
+        <div style="background-color: yellow; border: 1px solid purple; padding: 20px; display: flex; justify-content: center; align-items: center;">
+            <span style="color:red;">JavaScript is not enabled! This site won't work without it</span>
+        </div>
+    </noscript>
     <div class="navbar" role="navigation">
         <div class="container">
             <div class="navbar-collapse collapse">
@@ -149,8 +154,8 @@
         <footer class="main">
             <div class="container">
                 <div class="row flex-space-between">
-                    <div class="col-md-5">
-                        2018 - 2019 &copy; Nimiq URL Shortener.
+                    <div id="year_copyright" class="col-md-5">
+                        2018 - Today &copy; ShortNIM
                     </div>
                     <div class="col-md-7 text-right">
                         <a href="https://safe.nimiq.com/#_request/NQ65GS91H8CSQFAN1EVSUK3GX7PLL9N1X4KC_" target="_blank" style="margin: auto"><img src="https://www.nimiq.com/accept-donations/img/donationBtnImg/orange-small.svg" alt="Donate NIM" style="position: relative;top: 22%;"/></a>
@@ -166,6 +171,9 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.1.1/socket.io.js"></script>
     <script src="https://unpkg.com/sweetalert/dist/sweetalert.min.js"></script>
     <script src="statistics.js" defer></script>
+    <script>
+        document.getElementById('year_copyright').innerHTML = `2018 - ${(new Date).getFullYear()} &copy; ShortNIM`
+    </script>
 </body>
 
 </html>
