+++ public/statistics.js
