--- server/now.json
@@ -1,8 +1,0 @@
-{
-    "version": 1,
-    "env": {
-        "endpoint": "@endpoint",
-        "custom": "@custom"
-    },
-    "alias": "albermonte"
-}
